+++ lib/Mail/DMARC/Base.pm
@@ -2,7 +2,7 @@ package Mail::DMARC::Base;
 use strict;
 use warnings;
 use feature 'signatures';
-no warnings 'experimental::signatures';    ## no critic (ProhibitNoWarnings)
+no warnings 'experimental::args_array_with_signatures';    ## no critic (ProhibitNoWarnings)
 
 our $VERSION = '2.20260621';