+++ lib/Mail/DMARC/Report/Aggregate/Record/Auth_Results/DKIM.pm
@@ -3,7 +3,7 @@ our $VERSION = '2.20260621';
 use strict;
 use warnings;
 use feature 'signatures';
-no warnings 'experimental::signatures';    ## no critic (ProhibitNoWarnings)
+no warnings 'experimental::args_array_with_signatures';    ## no critic (ProhibitNoWarnings)
 
 use Carp;