+++ lib/Mail/DMARC/Report/Sender.pm
@@ -3,7 +3,6 @@ package Mail::DMARC::Report::Sender;
 use strict;
 use warnings;
 use feature 'signatures';
-no warnings 'experimental::signatures';    ## no critic (ProhibitNoWarnings)
 use feature 'try';
 no warnings 'experimental::try';    ## no critic (ProhibitNoWarnings)