+++ lib/Mail/DMARC/Report/Store/SQL.pm
@@ -3,7 +3,6 @@ our $VERSION = '2.20260621';
 use strict;
 use warnings;
 use feature 'signatures';
-no warnings 'experimental::signatures';    ## no critic (ProhibitNoWarnings)
 
 use Carp;
 use Data::Dumper;